Rename kernel_util to binary_util
