+++ tfjs-backend-cpu/src/kernels/Add.ts
@@ -18,7 +18,7 @@
 import {Add, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/kernel_utils';
+import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/binary_utils';
 
 export const addImpl = createSimpleBinaryKernelImpl(((a, b) => a + b));
 export const addComplexImpl =
