+++ tfjs-backend-cpu/src/kernels/Div.ts
@@ -18,7 +18,7 @@
 import {Div, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const divImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => a / b);
