+++ tfjs-backend-cpu/src/kernels/Equal.ts
@@ -18,7 +18,7 @@
 import {Equal, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const equalImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => (a === b) ? 1 : 0);
