+++ tfjs-backend-cpu/src/kernels/FloorDiv.ts
@@ -18,7 +18,7 @@
 import {FloorDiv, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const floorDivImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => Math.floor(a / b));
