+++ tfjs-backend-cpu/src/kernels/Greater.ts
@@ -18,7 +18,7 @@
 import {Greater, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const greaterImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => (a > b) ? 1 : 0);
