+++ tfjs-backend-cpu/src/kernels/GreaterEqual.ts
@@ -18,7 +18,7 @@
 import {GreaterEqual, KernelConfig} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const greaterEqualImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => (a >= b) ? 1 : 0);
