+++ tfjs-backend-cpu/src/kernels/Less.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, Less} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const lessImpl =
     createSimpleBinaryKernelImpl((a: number, b: number) => (a < b) ? 1 : 0);
