+++ tfjs-backend-cpu/src/kernels/Multiply.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Multiply} from '@tensorflow/tfjs-core';
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/kernel_utils';
+import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/binary_utils';
 
 export const multiplyImpl =
     createSimpleBinaryKernelImpl(((aValue, bValue) => aValue * bValue));
