+++ tfjs-backend-cpu/src/kernels/SquaredDifference.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, SquaredDifference} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc} from '../utils/kernel_utils';
+import {binaryKernelFunc} from '../utils/binary_utils';
 
 export const squaredDifferenceImpl = createSimpleBinaryKernelImpl(((a, b) => {
   const diff = a - b;
