+++ tfjs-backend-cpu/src/kernels/Sub.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, Sub} from '@tensorflow/tfjs-core';
 
 import {createSimpleBinaryKernelImpl} from '../utils/binary_impl';
-import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/kernel_utils';
+import {binaryKernelFunc, createComplexBinaryKernelImpl} from '../utils/binary_utils';
 
 export const subImpl =
     createSimpleBinaryKernelImpl(((aValue, bValue) => aValue - bValue));
