+++ tfjs-backend-cpu/src/utils/binary_utils.ts
